Separate UnicodeString and Traits into own files
