+++ SlashGaming-Diablo-II-API/include/cxx/game_struct/d2_unicode_char.hpp
@@ -46,19 +46,9 @@
 #ifndef SGD2MAPI_CXX_GAME_STRUCT_D2_UNICODE_CHAR_HPP_
 #define SGD2MAPI_CXX_GAME_STRUCT_D2_UNICODE_CHAR_HPP_
 
-#include <string>
-#include <string_view>
-#include <memory>
-
 #include "../../dllexport_define.inc"
 
 namespace d2 {
-
-/*
- * Due to the nature of how this API must be implemented (information hiding
- * being necessary to prevent user-created errors), all string APIs cannot be
- * derived from the standard library's.
- */
 
 struct UnicodeChar;
 
@@ -89,180 +79,6 @@ class DLLEXPORT UnicodeChar_Wrapper {
   UnicodeChar_Wrapper(UnicodeChar* ptr) noexcept;
   UnicodeChar_Wrapper(const UnicodeChar* ptr) noexcept;
 };
-
-class DLLEXPORT UnicodeChar_Traits {
- public:
-  using value_type = UnicodeChar;
-
-  static void assign(value_type& r, const value_type& a) noexcept;
-  static value_type* assign(value_type* p, std::size_t count, const value_type& a);
-  static bool eq(const value_type& a, const value_type& b);
-  static bool lt(const value_type& a, const value_type& b);
-  static value_type* move(value_type* dest, const value_type* src, std::size_t count);
-  static value_type* copy(value_type* dest, const value_type* src, std::size_t count);
-  static int compare(const value_type* s1, const value_type* s2, std::size_t count);
-  static const UnicodeChar_Traits::value_type* find(
-      const value_type* p,
-      std::size_t count,
-      const value_type& ch
-  );
-};
-
-class DLLEXPORT UnicodeString_API {
- public:
-  using value_type = UnicodeChar;
-  using size_type = int;
-  using reference = value_type&;
-  using const_reference = const value_type&;
-  using pointer = value_type*;
-  using const_pointer = const value_type*;
-
-  UnicodeString_API();
-  UnicodeString_API(size_type size, const value_type& ch);
-  UnicodeString_API(const UnicodeString_API& str);
-  UnicodeString_API(const UnicodeString_API& str, size_type pos);
-  UnicodeString_API(
-      const UnicodeString_API& str,
-      size_type pos,
-      size_type count
-  );
-  UnicodeString_API(const value_type* str);
-  UnicodeString_API(const value_type* str, size_type count);
-  UnicodeString_API(UnicodeString_API&& str) noexcept;
-
-  virtual ~UnicodeString_API();
-
-  UnicodeString_API& operator=(const UnicodeString_API& str);
-  UnicodeString_API& operator=(UnicodeString_API&& other) noexcept;
-  UnicodeString_API& operator=(const value_type* str);
-  UnicodeString_API& operator=(const value_type& str);
-
-  UnicodeString_API& operator+=(const UnicodeString_API& str);
-  UnicodeString_API& operator+=(const value_type& ch);
-  UnicodeString_API& operator+=(const value_type* str);
-
-  reference operator[](size_type pos);
-  const_reference operator[](size_type pos) const;
-
-  static UnicodeString_API FromU8String(std::u8string_view src);
-
-  std::u8string ToU8String() const;
-
-  UnicodeString_API& append(size_type count, const value_type& ch);
-  UnicodeString_API& append(const UnicodeString_API& str);
-  UnicodeString_API& append(
-      const UnicodeString_API& str,
-      size_type pos
-  );
-  UnicodeString_API& append(
-      const UnicodeString_API& str,
-      size_type pos,
-      size_type count
-  );
-  UnicodeString_API& append(const value_type* str);
-  UnicodeString_API& append(const value_type* str, size_type count);
-
-  reference at(size_type pos);
-  const_reference at(size_type pos) const;
-
-  value_type& back();
-  const value_type& back() const;
-
-  void clear() noexcept;
-
-  size_type copy(
-      value_type* dest,
-      size_type count
-  ) const;
-  size_type copy(
-      value_type* dest,
-      size_type count,
-      size_type pos
-  ) const;
-
-  [[nodiscard]]
-  bool empty() const noexcept;
-
-  value_type& front();
-  const value_type& front() const;
-
-  void pop_back();
-
-  void push_back(const value_type& ch);
-
-  void resize(size_type count);
-  void resize(size_type count, const value_type& ch);
-
-  UnicodeString_API substr() const;
-  UnicodeString_API substr(size_type pos) const;
-  UnicodeString_API substr(size_type pos, size_type count) const;
-
-  void swap(UnicodeString_API& other);
-
-  UnicodeChar* data() noexcept;
-  const UnicodeChar* data() const noexcept;
-
-  size_type capacity() const noexcept;
-  const UnicodeChar* c_str() const noexcept;
-  size_type size() const noexcept;
-  size_type length() const noexcept;
-
- private:
-  UnicodeChar* data_;
-  size_type capacity_;
-  size_type length_;
-
-  static void TerminateString(value_type* str, size_type pos);
-};
-
-UnicodeString_API operator+(
-    const UnicodeString_API& lhs,
-    const UnicodeString_API& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeString_API& lhs,
-    const UnicodeChar* rhs
-);
-UnicodeString_API operator+(
-    const UnicodeString_API& lhs,
-    const UnicodeChar& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeChar* lhs,
-    const UnicodeString_API& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeChar& lhs,
-    const UnicodeString_API& rhs
-);
-UnicodeString_API operator+(
-    UnicodeString_API&& lhs,
-    UnicodeString_API&& rhs
-);
-UnicodeString_API operator+(
-    UnicodeString_API&& lhs,
-    const UnicodeString_API& rhs
-);
-UnicodeString_API operator+(
-    UnicodeString_API&& lhs,
-    const UnicodeChar* rhs
-);
-UnicodeString_API operator+(
-    UnicodeString_API&& lhs,
-    const UnicodeChar& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeString_API& lhs,
-    UnicodeString_API&& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeChar* lhs,
-    UnicodeString_API&& rhs
-);
-UnicodeString_API operator+(
-    const UnicodeChar& lhs,
-    UnicodeString_API&& rhs
-);
 
 } // namespace d2
