+++ SlashGaming-Diablo-II-API/src/cxx/game_struct/d2_unicode_char/d2_unicode_string_api.cc
@@ -43,7 +43,7 @@
  *  work.
  */
 
-#include "../../../../include/cxx/game_struct/d2_unicode_char.hpp"
+#include "../../../../include/cxx/game_struct/d2_unicode_string_api.hpp"
 
 #include <algorithm>
 #include <stdexcept>
@@ -51,6 +51,7 @@
 #include "../../../../include/cxx/game_func/d2lang_func.hpp"
 #include "../../../../include/cxx/game_version.hpp"
 #include "d2_unicode_char_impl.hpp"
+#include "../../../../include/cxx/game_struct/d2_unicode_char_traits_api.hpp"
 
 namespace d2 {
 namespace {
